+++ internal/cmd/notimplemented_generator/standardops.go
@@ -9,6 +9,7 @@ import (
 	"text/template"
 
 	"fmt"
+
 	"github.com/gomlx/gomlx/internal/cmd/backends_generator/parsexlabuilder"
 	"github.com/gomlx/gomlx/types"
 	"github.com/gomlx/gomlx/types/xslices"
@@ -35,13 +36,12 @@ import (
 	"github.com/gomlx/gomlx/backends"
 	"github.com/gomlx/gomlx/types/shapes"
 	"github.com/gomlx/gopjrt/dtypes"
-	"github.com/pkg/errors"
 )
 
 {{range .}}
 {{range .Comments}}// {{.}}
 {{end}}func (b Builder) {{.Name}}({{range .Parameters}}{{.Names}} {{.Type}},{{end}}) (backends.Op, error) {
-	return nil, errors.Wrapf(NotImplementedError, "in method {{.Name}}()")
+	return nil, b.baseErrFn(backends.OpType{{.Name}})
 }
 {{end}}
 `))